splits SDKConnect into two classes in different files
